Stick footer to top and bottom
Create a flex-container of the 'body' element
and orient it in a column with space between.

Create a flex-container for the middle content of
the page to orient in a row. Cause this container
to grow and fill the 'space-between' created by
the flex properties defined in the 'body' element.